Fix daylight savings test failures (#49)
Before, the tests used the "Europe/London" timezone.
However, this caused issues because when daylight
savings occurred (British Summer Time) all of the
times (the hour) would be off by one.

Now, we update the tests to use the "UTC" timezone
which does not change during the year.

Later, I should probably refactor the tests to be
fixed in time rather than dynamic from today. But
this fix will do for now!