Fix blank page in outlook.com after saving event (#56)
Before, visiting the outlook_com_url event link would display the event
form to save, but once saved you would see a blank screen.

Now, we update the URL so that the event form opens in-calendar, and
saving the event returns you to the calendar view as expected.

This change is all thanks to @CobiOneCanobi who spotted the issue and
submitted a fix for Office365. Thank you!